backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
thefuzz
python-Levenshtein
